+++ common/go.mod
@@ -1,0 +1,8 @@
+module github.com/jessemillman/demo-confluence-reporting/common
+
+go 1.12
+
+require (
+	github.com/gocarina/gocsv v0.0.0-20200330101823-46266ca37bd3
+	github.com/pkg/errors v0.9.1
+)
