+++ config/go.mod
@@ -1,0 +1,5 @@
+module github.com/jessemillman/demo-confluence-reporting/config
+
+go 1.12
+
+require github.com/pkg/errors v0.9.1
